requirements.txt
Flask==2.2.5
langchain==0.3.1  # Same as Colab
chromadb==0.5.11
chroma-hnswlib==0.7.6
google-generativeai==0.7.2  # Same as Colab
pypdf==5.0.0
numpy==1.26.4
pandas==2.1.4
requests==2.32.3
transformers==4.44.2
torch==2.0.1  # Compatible for Mac with CPU
openai==0.27.7
scikit-learn==1.5.2
tqdm==4.66.5
Jinja2==3.1.4
Werkzeug==3.0.4
itsdangerous==2.2.0
streamlit
langchain-community==0.3.1
fastapi==0.115.0
PyMuPDF==1.24.10
beautifulsoup4==4.12.3
langchain-google-genai==2.0.0  # Add this line to include the module